Move google login button to Navbar.
